move metafunc tests in their own file
