+++ spy/tests/compiler/test_basic.py
@@ -1094,31 +1094,6 @@ class TestBasic(CompilerTest):
         """)
         assert mod.foo(3) == 4
 
-    def test_blue_metafunc(self):
-        mod = self.compile("""
-        from operator import OpSpec
-
-        @blue.metafunc
-        def foo(v_x):
-            if v_x.static_type == i32:
-               def impl_i32(x: i32) -> i32:
-                   return x * 2
-               return OpSpec(impl_i32)
-            elif v_x.static_type == str:
-               def impl_str(x: str) -> str:
-                   return x + ' world'
-               return OpSpec(impl_str)
-            raise StaticError("unsupported type")
-
-        def test1() -> i32:
-            return foo(5)
-
-        def test2() -> str:
-            return foo('hello')
-        """)
-        assert mod.test1() == 10
-        assert mod.test2() == 'hello world'
-
     def test_call_module_attr(self):
         mod = self.compile("""
         import math
